+++ routes/web.php
@@ -1,5 +1,0 @@
-<?php
-
-// @deprecated
-Route::get('!/statamic-lazy-logo.svg', 'CP\LogoController');
-Route::get('!/lazy-logo.svg', 'CP\LogoController');