+++ src/ServiceProvider.php
@@ -7,8 +7,7 @@ use Statamic\Providers\AddonServiceProvider;
 class ServiceProvider extends AddonServiceProvider
 {
     protected $routes = [
-        'web' => __DIR__.'/../routes/web.php',
-        'cp'  => __DIR__.'/../routes/cp.php',
+        'cp' => __DIR__.'/../routes/cp.php',
     ];
 
     protected $updateScripts = [