+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.0 (2022-12-12)
+
+- [new] Supports different logo styles for outside (login etc.) and inside pages
+- [new] Supports setting an outside page background colour or gradient
+- [fix] Image and style browser caching
+- [break] The `custom_logo_url` value has changed, see readme
+- Deprecated routes have been removed
+
 ## 0.3.3 (2021-12-13)
 
 - [fix] Logo route should not require authentication